pyproject.toml
[tool.poetry]
name = "zoneh"
version = "0.1.0"
description = ""
authors = ["Ahmed Elshahat <58784045+MrStarkEG@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
playwright = "^1.49.1"
python-dotenv = "^1.0.1"
anticaptchaofficial = "^1.0.62"
requests = "^2.32.3"
tenacity = "^9.0.0"
bs4 = "^0.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
